requirements.txt
alabaster==0.7.12
amqp==5.0.6
anyio==3.3.4
argon2-cffi==21.1.0
asgiref==3.4.1
astroid==2.8.4
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
black==21.9b0
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
colorlog==6.5.0
coreapi==2.3.3
coreschema==0.0.4
coverage==6.0.2
cryptography==35.0.0
Cython==0.29.24
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.3
Django==3.1.13
django-allauth==0.45.0
django-appconf==1.0.5
django-celery-beat==2.2.1
django-compressor==2.4.1
django-cors-headers==3.10.0
django-coverage-plugin==2.0.1
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-environ==0.7.0
django-extensions==3.1.3
django-model-utils==4.2.0
django-redis==5.0.0
django-rq==2.4.1
django-stubs==1.8.0
django-stubs-ext==0.3.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-stubs==1.4.0
docutils==0.17.1
EasyProcess==0.3
entrypoint2==0.2.4
factory-boy==3.2.0
Faker==9.7.1
filelock==3.3.1
flake8==4.0.1
flake8-isort==4.1.1
flower==1.0.0
h11==0.12.0
hiredis==2.0.0
httptools==0.2.0
humanize==3.12.0
identify==2.3.1
idna==3.3
imagesize==1.2.0
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.28.0
iso-639==0.4.5
iso3166==2.0.2
isodate==0.6.0
isort==5.9.3
itypes==1.2.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.2
kombu==5.1.0
lazy-object-proxy==1.6.0
livereload==2.6.3
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
msgpack==1.0.2
msgpack-numpy==0.4.7.1
mss==6.1.0
mypy==0.910
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.21.3
oauthlib==3.1.1
opencv-contrib-python-headless==4.5.3.56
packaging==21.0
pafy==0.5.5
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.4.0
pluggy==1.0.0
pre-commit==2.15.0
prometheus-client==0.11.0
prompt-toolkit==3.0.21
psycopg2==2.9.1
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.8.0
pycparser==2.20
pycryptodome==3.11.0
pyflakes==2.4.0
Pygments==2.10.0
PyJWT==2.3.0
pylint==2.11.1
pylint-celery==0.3
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==3.0.3
pyscreenshot==3.0
PySocks==1.7.1
pytest==6.2.5
pytest-django==4.4.0
pytest-sugar==0.9.4
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.1
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
rcssmin==1.0.6
redis==3.5.3
regex==2021.10.23
requests==2.26.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
rq==1.10.0
simplejpeg==1.6.2
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.1.0
Sphinx==4.2.0
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.2
starlette==0.16.0
streamlink==2.4.0
termcolor==1.1.0
testfixtures==6.18.3
text-unidecode==1.3
toml==0.10.2
tomli==1.2.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
vidgear==0.2.2
vine==5.0.0
virtualenv==20.9.0
watchgod==0.7
wcwidth==0.2.5
websocket-client==1.2.1
websockets==10.0
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.13.2
youtube-dl==2021.6.6
